Added ThermoVacuum Testing Program for NASA Test
There are 2 projects now under the experimental folder: ThermoVacDummyBoard and ThermoVacTestBoard. ThermoVacTestBoard is programmed to the prototype that will be connected to the FTDI cable and thus PC for recording throughput and packet error rate measurements. Dummy board is for the other board so the Test board has something to send data to.